Update LinuxMain. More test refactoring.
