+++ Sources/Redis+Geo.swift
@@ -167,14 +167,10 @@ extension Redis {
         var command = ["GEODIST", key, member1, member2]
         if let unit = unit {
             switch unit {
-            case .m:
-                command.append(GeoDistanceUnit.m.rawValue)
-            case .km:
-                command.append(GeoDistanceUnit.km.rawValue)
-            case .mi:
-                command.append(GeoDistanceUnit.mi.rawValue)
-            case .ft:
-                command.append(GeoDistanceUnit.ft.rawValue)
+            case .m: command.append(GeoDistanceUnit.m.rawValue)
+            case .km: command.append(GeoDistanceUnit.km.rawValue)
+            case .mi: command.append(GeoDistanceUnit.mi.rawValue)
+            case .ft: command.append(GeoDistanceUnit.ft.rawValue)
             }
         }
         issueCommandInArray(command) { (response) in
@@ -225,14 +221,10 @@ extension Redis {
     public func georadius(key: String, longitude: Double, latitude: Double, radius: Double, unit: GeoDistanceUnit, withCoord: Bool?=nil, withDist: Bool?=nil, withHash: Bool?=nil, count: Int?=nil, ascending: Bool?=nil, callback: (_ result: [RedisResponse?]?, _ error: NSError?) -> Void) {
         var command = ["GEORADIUS", key, String(longitude), String(latitude), String(radius)]
         switch unit {
-        case .m:
-            command.append(GeoDistanceUnit.m.rawValue)
-        case .km:
-            command.append(GeoDistanceUnit.km.rawValue)
-        case .mi:
-            command.append(GeoDistanceUnit.mi.rawValue)
-        case .ft:
-            command.append(GeoDistanceUnit.ft.rawValue)
+        case .m: command.append(GeoDistanceUnit.m.rawValue)
+        case .km: command.append(GeoDistanceUnit.km.rawValue)
+        case .mi: command.append(GeoDistanceUnit.mi.rawValue)
+        case .ft: command.append(GeoDistanceUnit.ft.rawValue)
         }
         if let withCoord = withCoord, withCoord {
             command.append("WITHCOORD")
@@ -300,14 +292,10 @@ extension Redis {
     public func georadiusbymember(key: String, member: String, radius: Double, unit: GeoDistanceUnit, withCoord: Bool?=nil, withDist: Bool?=nil, withHash: Bool?=nil, count: Int?=nil, ascending: Bool?=nil, callback: (_ result: [RedisResponse?]?, _ error: NSError?) -> Void) {
         var command = ["GEORADIUSBYMEMBER", key, member, String(radius)]
         switch unit {
-        case .m:
-            command.append(GeoDistanceUnit.m.rawValue)
-        case .km:
-            command.append(GeoDistanceUnit.km.rawValue)
-        case .mi:
-            command.append(GeoDistanceUnit.mi.rawValue)
-        case .ft:
-            command.append(GeoDistanceUnit.ft.rawValue)
+        case .m: command.append(GeoDistanceUnit.m.rawValue)
+        case .km: command.append(GeoDistanceUnit.km.rawValue)
+        case .mi: command.append(GeoDistanceUnit.mi.rawValue)
+        case .ft: command.append(GeoDistanceUnit.ft.rawValue)
         }
         if let withCoord = withCoord, withCoord {
             command.append("WITHCOORD")
