+++ Sources/RedisMulti+Geo.swift
@@ -22,6 +22,11 @@ extension RedisMulti {
     /// (longitude, latitude, name)
     public typealias GeospatialItem = (Double, Double, String)
     
+    /// Unit of distance used in various GEO commands
+    public enum GeoDistanceUnit: String {
+        case m, km, mi, ft
+    }
+    
     /// Adds the specified geospatial items (longitude, latitude, name) to the
     /// specified key.
     ///
@@ -133,10 +138,15 @@ extension RedisMulti {
     ///
     /// - Returns: The `RedisMulti` object being added to.
     @discardableResult
-    public func geodist(key: String, member1: String, member2: String, unit: String?=nil) -> RedisMulti {
+    public func geodist(key: String, member1: String, member2: String, unit: GeoDistanceUnit?=nil) -> RedisMulti {
         var command = ["GEODIST", key, member1, member2]
         if let unit = unit {
-            command.append(unit)
+            switch unit {
+            case .m: command.append(GeoDistanceUnit.m.rawValue)
+            case .km: command.append(GeoDistanceUnit.km.rawValue)
+            case .mi: command.append(GeoDistanceUnit.mi.rawValue)
+            case .ft: command.append(GeoDistanceUnit.ft.rawValue)
+            }
         }
         queuedCommands.append(stringArrToRedisStringArr(command))
         return self
@@ -179,8 +189,14 @@ extension RedisMulti {
     ///
     /// - Returns: The `RedisMulti` object being added to.
     @discardableResult
-    public func georadius(key: String, longitude: Double, latitude: Double, radius: Double, unit: String, withCoord: Bool?=nil, withDist: Bool?=nil, withHash: Bool?=nil, count: Int?=nil, ascending: Bool?=nil) -> RedisMulti {
-        var command = ["GEORADIUS", key, String(longitude), String(latitude), String(radius), unit]
+    public func georadius(key: String, longitude: Double, latitude: Double, radius: Double, unit: GeoDistanceUnit, withCoord: Bool?=nil, withDist: Bool?=nil, withHash: Bool?=nil, count: Int?=nil, ascending: Bool?=nil) -> RedisMulti {
+        var command = ["GEORADIUS", key, String(longitude), String(latitude), String(radius)]
+        switch unit {
+        case .m: command.append(GeoDistanceUnit.m.rawValue)
+        case .km: command.append(GeoDistanceUnit.km.rawValue)
+        case .mi: command.append(GeoDistanceUnit.mi.rawValue)
+        case .ft: command.append(GeoDistanceUnit.ft.rawValue)
+        }
         if let withCoord = withCoord, withCoord {
             command.append("WITHCOORD")
         }
@@ -191,6 +207,7 @@ extension RedisMulti {
             command.append("WITHHASH")
         }
         if let count = count {
+            command.append("COUNT")
             command.append(String(count))
         }
         if let ascending = ascending {
@@ -239,8 +256,14 @@ extension RedisMulti {
     ///
     /// - Returns: The `RedisMulti` object being added to.
     @discardableResult
-    public func georadiusbymember(key: String, member: String, radius: Double, unit: String, withCoord: Bool?=nil, withDist: Bool?=nil, withHash: Bool?=nil, count: Int?=nil, ascending: Bool?=nil) -> RedisMulti {
-        var command = ["GEORADIUSBYMEMBER", key, member, String(radius), unit]
+    public func georadiusbymember(key: String, member: String, radius: Double, unit: GeoDistanceUnit, withCoord: Bool?=nil, withDist: Bool?=nil, withHash: Bool?=nil, count: Int?=nil, ascending: Bool?=nil) -> RedisMulti {
+        var command = ["GEORADIUSBYMEMBER", key, member, String(radius)]
+        switch unit {
+        case .m: command.append(GeoDistanceUnit.m.rawValue)
+        case .km: command.append(GeoDistanceUnit.km.rawValue)
+        case .mi: command.append(GeoDistanceUnit.mi.rawValue)
+        case .ft: command.append(GeoDistanceUnit.ft.rawValue)
+        }
         if let withCoord = withCoord, withCoord {
             command.append("WITHCOORD")
         }
@@ -251,6 +274,7 @@ extension RedisMulti {
             command.append("WITHHASH")
         }
         if let count = count {
+            command.append("COUNT")
             command.append(String(count))
         }
         if let ascending = ascending {
